+++ tapeserver/castor/tape/tapeserver/drive/DriveInterface.hpp
@@ -172,8 +172,8 @@ namespace drive {
     
     virtual compressionStats getCompression()  = 0;
     virtual void clearCompressionStats()  = 0;
-    virtual std::map<std::string,uint32_t> getTapeWriteErrors() = 0;
-    virtual std::map<std::string,uint32_t> getTapeReadErrors() = 0;
+    virtual std::map<std::string,uint64_t> getTapeWriteErrors() = 0;
+    virtual std::map<std::string,uint64_t> getTapeReadErrors() = 0;
     virtual std::map<std::string,uint32_t> getTapeNonMediumErrors() = 0;
     virtual std::map<std::string,float> getQualityStats() = 0;
     virtual std::map<std::string,uint32_t> getDriveStats() = 0;
